Fixed cell-indexing
