+++ Jottre/Controllers/SettingsScene/SettingsExtensions.swift
@@ -40,7 +40,6 @@ extension SettingsViewController: UICollectionViewDelegate, UICollectionViewData
             guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "appearanceSettingsCell", for: indexPath) as? AppearanceSettingsCell else {
                 fatalError("Cell is not of type AppearanceSettingsCell")
             }
-            cell.title = NSLocalizedString("Interface appearance", comment: "")
             return cell
             
         } else if indexPath.row == 1 {
@@ -49,11 +48,7 @@ extension SettingsViewController: UICollectionViewDelegate, UICollectionViewData
                 fatalError("Cell is not of type CloudSettingsCell")
             }
             cell.delegate = self
-            cell.title = NSLocalizedString("Synchronize with iCloud", comment: "")
             cell.usesCloud = UIDevice.isLimited() ? true : settings.codable.usesCloud
-            if UIDevice.isLimited() || !Downloader.isCloudEnabled {
-                cell.switchView.isEnabled = false
-            }
             return cell
             
         } else if indexPath.row == 2 {
